cambios tipo de dato ID de integer a Long

+++ academic/academic/src/main/java/edu/unimagdalena/academic/repositories/AsignaturaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import edu.unimagdalena.academic.entities.Asignatura;
 
 @Repository
-public interface AsignaturaRepository extends JpaRepository<Asignatura, Integer>{
+public interface AsignaturaRepository extends JpaRepository<Asignatura, Long>{
 
 }
