+++ academic/academic/src/main/java/edu/unimagdalena/academic/repositories/EstudianteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import edu.unimagdalena.academic.entities.Estudiante;
 
 @Repository
-public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
 
 }
