+++ academic/academic/src/main/java/edu/unimagdalena/academic/repositories/Hora_SemanalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import edu.unimagdalena.academic.entities.Hora_Semanal;
 @Repository
-public interface Hora_SemanalRepository extends JpaRepository< Hora_Semanal, Integer> {
+public interface Hora_SemanalRepository extends JpaRepository< Hora_Semanal, Long> {
 
 }
