+++ academic/academic/src/main/java/edu/unimagdalena/academic/repositories/RoleRepository.java
@@ -3,6 +3,6 @@ package edu.unimagdalena.academic.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import edu.unimagdalena.academic.entities.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Integer>{
+public interface RoleRepository extends JpaRepository<Role, Long>{
 
 }
